% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agribbit.R
\name{agri.fast_draw}
\alias{agri.fast_draw}
\title{to interpolate missing values}
\usage{
agri.fast_draw(
  df,
  variable,
  xlab = "x",
  ylab = "y",
  fill = "",
  fill_low = "cyan",
  fill_high = "tomato"
)
}
\description{
\code{agri.fast_draw}
}
