% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agribbit.R
\name{agri.interpolate}
\alias{agri.interpolate}
\title{to interpolate missing values}
\usage{
agri.interpolate(df, obj, kernel = "rbfdot")
}
\arguments{
\item{引数}{説明}
}
\value{
出力の説明
}
\description{
\code{agri.interpolate}
}
\examples{
関数の例

}
