% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agribbit.R
\name{agri.sf_plot_continuous}
\alias{agri.sf_plot_continuous}
\title{to interpolate missing values}
\usage{
agri.sf_plot_continuous(df, variable, xlab = "x", ylab = "y", fill = "")
}
\description{
\code{agri.sf_plot_continuous}
}
